ForwardTacotron-master/requirements.txt
numba==0.49.1
librosa==0.7.2
pyworld >= 0.2.10
torch>=1.2.0
phonemizer>=2.2
webrtcvad>=2.0.10
PyYAML>=5.1
soundfile
scipy
tensorboard
matplotlib
unidecode
inflect
